+++ atomicshop/basics/lists.py
@@ -1,4 +1,3 @@
-# v1.0.4 - 26.03.2023 - 12:50
 def remove_duplicates(list_instance: list):
     # One of the fastest methods.
     seen = set()
@@ -24,3 +23,22 @@ def get_difference(list1, list2):
     # This is one of the fastest approaches: https://stackoverflow.com/a/3462202
     s = set(list2)
     return [x for x in list1 if x not in s]
+
+
+def replace_elements_with_values_from_dict(list_instance: list, dictionary: dict) -> list:
+    """
+    Function exchanges elements from list with elements from dictionary.
+
+    Example:
+        list_instance = ['a', 'b', 'c']
+        dictionary = {'a': 'd', 'b': 'e', 'c': 'f'}
+        exchange_elements_from_dict(list_instance, dictionary)
+    Result:
+        ['d', 'e', 'f']
+
+    :param list_instance: list.
+    :param dictionary: dictionary.
+    :return: list.
+    """
+
+    return [dictionary.get(x, x) for x in list_instance]